Better document the audio primitives and abstractions. This does also update the WaveStream base class documentation, as well as the IWavePlayer docs. The IWavePlayer declaration is also cleaned up and it's extension interfaces are now provided in separate files. Additionally, this fixes a minor bounding error in WaveStream, that it's CurrentTime property could set a Position value way larger than Length, or way smaller than 0. Finally, the WaveProviders doc is updated to reflect the new changes that were performed in NAudio 3.
