+++ src/NAudio.Core/Wave/WaveOutputs/IWavePlayer.cs
@@ -1,124 +1,103 @@
-﻿using System;
+﻿
+using System;
 
 // ReSharper disable once CheckNamespace
 namespace NAudio.Wave;
 
 /// <summary>
-/// Represents the interface to a device that can play a WaveFile
+/// Represents a hardware audio device that can play any <see cref="IWaveProvider"/> 
+/// instance provided to it's <see cref="Init" /> method. <br />
+/// This interface is typically implemented by audio drivers, such as WASAPI on Windows. <br />
+/// You are expected to release the acquired audio hardware 
+/// (by calling <see cref="IDisposable.Dispose" />) once you are finished using it.
 /// </summary>
+/// <seealso cref="IWavePosition" />
+/// <seealso cref="IWaveLatency" />
 public interface IWavePlayer : IDisposable
 {
-    /// <summary>
-    /// Begin playback
-    /// </summary>
+    /// <summary>Begin playback</summary>
     void Play();
 
-    /// <summary>
-    /// Stop playback
-    /// </summary>
+    /// <summary>Stop playback</summary>
     void Stop();
 
-    /// <summary>
-    /// Pause Playback
-    /// </summary>
+    /// <summary>Pause Playback</summary>
+    /// <remarks>
+    /// The way this method works is that during the Pause state,
+    /// the buffers of the driver are not yet cleared, but the
+    /// playback has been stopped. As such, later resuming playback
+    /// is faster than doing <see cref="Stop"/>, then <see cref="Play"/>. <br />
+    /// Note that not all drivers do necessarily implement this
+    /// because it depends on the programming model the driver is using.
+    /// </remarks>
     void Pause();
 
     /// <summary>
-    /// Initialise playback
+    /// Initialises playback from a given wave
+    /// provider that provides the samples to play.
     /// </summary>
-    /// <param name="waveProvider">The wave provider to be played</param>
+    /// <param name="waveProvider">The wave provider to be played.</param>
+    /// <exception cref="ArgumentNullException"><paramref name="waveProvider" /> is <see langword="null" />.</exception>
     void Init(IWaveProvider waveProvider);
 
     /// <summary>
-    /// The volume 
-    /// 1.0f is full scale
-    /// Note that not all implementations necessarily support volume changes
+    /// Allows to modify the gain (volume) of the output audio. <br />
+    /// 1.0f is full scale, while 0.0f is silence.
     /// </summary>
+    /// <returns>
+    /// A value indicating the gain (volume) of the output audio. <br />
+    /// Implementations are permitted to always return 1f if the driver does
+    /// not have gain controls.
+    /// </returns>
+    /// <remarks>
+    /// Not all implementations do necessarily support getting/setting a volume value. <br />
+    /// In some drivers, this modifies the stream's volume, and in some others the global or the device's hardware volume.
+    /// </remarks>
+    /// <exception cref="InvalidOperationException">When setting the property: Volume changes are not supported.</exception>
+    /// <exception cref="ArgumentOutOfRangeException">When setting the property: The new volume exceeds 1f, or is less than 0f.</exception>
     float Volume { get; set; }
 
     /// <summary>
-    /// Current playback state
+    /// Queries the current state of the playback.
     /// </summary>
+    /// <returns>
+    /// A value from the <see cref="Wave.PlaybackState"/> 
+    /// enumeration that indicates the state of the playback.
+    /// </returns>
     PlaybackState PlaybackState { get; }
 
     /// <summary>
-    /// Indicates that playback has gone into a stopped state due to 
-    /// reaching the end of the input stream or an error has been encountered during playback
+    /// The audio format under which the driver
+    /// provides samples to the connected-to hardware device.
+    /// </summary>
+    /// <remarks>
+    /// This is not necessarily the same format as the 
+    /// one provided from the <see cref="IWaveProvider" /> during <see cref="Init" />.
+    /// </remarks>
+    WaveFormat OutputWaveFormat { get; }
+
+    /// <summary>
+    /// Event that is dispatching when playback has been stopped. <br />
+    /// Playback can be stopped due to any of the following three reasons: <br />
+    /// <list type="bullet">
+    ///     <item>
+    ///         You called the <see cref="Stop"/> method. <br />
+    ///         In this case, the event args object 
+    ///         <see cref="StoppedEventArgs.Exception" /> property is <see langword="null" />.
+    ///     </item>
+    ///     <item>
+    ///         The provided <see cref="IWaveProvider"/> instance
+    ///         has reached it's end (no more samples to play). <br />
+    ///         In this case, the event args object 
+    ///         <see cref="StoppedEventArgs.Exception" /> property is <see langword="null" />.
+    ///     </item>
+    ///     <item>
+    ///         An error/exception has been occurred either by calling <see cref="IWaveProvider.Read(Span{byte})" />,
+    ///         or the driver has encountered a failure which it forced to stop playback. <br />
+    ///         In this case, the event args object 
+    ///         <see cref="StoppedEventArgs.Exception" /> property is <em>not</em> <see langword="null" />.
+    ///     </item>
+    /// </list>
     /// </summary>
     event EventHandler<StoppedEventArgs> PlaybackStopped;
-
-    /// <summary>
-    /// The WaveFormat this device is using for playback
-    /// </summary>
-    WaveFormat OutputWaveFormat { get; }
-}
-
-/// <summary>
-/// Interface for IWavePlayers that can report position
-/// </summary>
-public interface IWavePosition
-{
-    /// <summary>
-    /// Position (in terms of bytes played - does not necessarily translate directly to the position within the source audio file)
-    /// </summary>
-    /// <returns>Position in bytes</returns>
-    long GetPosition();
-
-    /// <summary>
-    /// Gets a <see cref="Wave.WaveFormat"/> instance indicating the format the hardware is using.
-    /// </summary>
-    WaveFormat OutputWaveFormat { get; }
-}
-
-/// <summary>
-/// Interface for audio players and captures that can report the latency between a sample
-/// entering the audio pipeline and emerging from (or being delivered by) the hardware.
-/// Implement alongside <see cref="IWavePlayer"/> or <see cref="IWaveIn"/> to let downstream
-/// code synchronise visualisations, lighting, video, recording timecode, or any other
-/// time-aligned consumer with audible output or captured input.
-/// </summary>
-/// <remarks>
-/// <para>
-/// Latency values are only meaningful during uninterrupted operation. When the device is
-/// stopped or paused, or when an underrun has occurred, implementations should return their
-/// best steady-state estimate rather than throw — consumers driving visualisations need a
-/// stable value to fall back on.
-/// </para>
-/// <para>
-/// The two properties answer different questions. <see cref="AverageLatency"/> describes the
-/// pipeline (it changes only when buffer settings change), so it is the right value for
-/// scheduling. <see cref="CurrentLatency"/> describes the live state of the pipeline at this
-/// instant and is the right value for drift detection or per-frame correction.
-/// </para>
-/// <para>
-/// <see cref="CurrentLatency"/> may be computed either as the forward queue depth (e.g. WASAPI
-/// <c>GetCurrentPadding</c> divided by sample rate — "if I queued a sample now, this is how
-/// long until I'd hear it") or as the wall-clock age of the sample currently at the play /
-/// capture head (timestamping each buffer fill). The two values converge in steady state and
-/// diverge only under irregular feed patterns; this interface deliberately permits either,
-/// so implementations can use whichever the underlying driver exposes most cheaply.
-/// </para>
-/// <para>
-/// For drivers whose buffer scheduling is fully predictable (notably ASIO, which always swaps
-/// fixed-size buffers at regular intervals), <see cref="CurrentLatency"/> may simply return
-/// <see cref="AverageLatency"/>. The approximation is exact to within half a buffer.
-/// </para>
-/// </remarks>
-public interface IWaveLatency
-{
-    /// <summary>
-    /// The steady-state latency from a sample being queued for output to it being emitted by
-    /// the audio hardware, assuming uninterrupted playback. This is a property of the buffer
-    /// configuration and the driver, not of the current playback state.
-    /// </summary>
-    TimeSpan AverageLatency { get; }
-
-    /// <summary>
-    /// The time that has elapsed since the sample currently emerging from the hardware was
-    /// queued for output. In steady state this is approximately equal to
-    /// <see cref="AverageLatency"/>; it differs during start-up, after an underrun, or when
-    /// the host is filling buffers irregularly. Implementations that cannot meaningfully
-    /// distinguish from the average are permitted to return <see cref="AverageLatency"/>.
-    /// </summary>
-    TimeSpan CurrentLatency { get; }
 }
